requirements.txt
aiohttp==3.7.3
astor==0.8.1
async-timeout==3.0.1
attrs==20.2.0
autocorrect==2.1.0
beautifulsoup4==4.9.2
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Flask==1.0.2
futures==3.1.1
goslate==1.5.1
idna==2.10
idna-ssl==1.1.0
itsdangerous==0.24
Jinja2==2.10
joblib==0.16.0
lark-parser==0.10.0
linereader==1.0.0
multidict==4.7.6
nltk==3.5
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.2
py2exe==0.9.2.0
PyDictionary==2.0.1
pyLibrary==3.47.20042
pyparsing==2.2.0
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2017.3
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
schedule==0.6.0
scikit-learn==0.20.0
scipy==1.1.0
six==1.15.0
sklearn==0.0
slack==0.0.2
slack-client==0.3.0
slack-webclient-logger==1.0.4
slackclient==2.9.1
soupsieve==2.0.1
tqdm==4.49.0
tweepy==3.9.0
typing-extensions==3.7.4.3
tython==0.0.1a0
urllib3==1.25.10
websocket-client==0.53.0
Werkzeug==0.14.1
yarl==1.6.0
